requirements.txt
pydantic~=2.10.4
torch~=2.5.1
transformers~=4.47.1
PyYAML~=6.0.2
pdfplumber~=0.11.4
Markdown~=3.7
beautifulsoup4~=4.12.3
python-dotenv~=1.0.1
langchain~=0.3.13
pytest
pytest-cov
langchain_community
sentence-transformers
faiss-cpu
pyyaml
pydantic
python-dotenv
pdfplumber
beautifulsoup4
markdown
